django_server/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
drf-yasg==1.20.0
gunicorn==20.0.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pymongo==3.12.0
pyparsing==2.4.7
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.2.4
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
